add innitial candle webgpu kernels as well as dummy backend

+++ candle-core/src/device.rs
@@ -9,7 +9,7 @@ pub enum DeviceLocation {
     Cpu,
     Cuda { gpu_id: usize },
     Metal { gpu_id: usize },
-    WebGPU { gpu_id: usize },
+    WebGPU { gpu_id: u64 },
 }
 
 #[derive(Debug, Clone)]
